+++ addons/filamind_kitchen_display/models/__init__.py
@@ -1,0 +1,6 @@
+from . import filamind_kitchen_display
+from . import filamind_kitchen_stage
+from . import filamind_kitchen_order
+from . import filamind_kitchen_line
+from . import pos_config
+from . import pos_order